The TODO list keeps growing.
But I'm happy because this tool is now the MVP, the minimum viable
product. It has all the features I needed to make it useful for me. I
can change individual volumes, and I can change the display brightness.
In other words, I can start using this tool right away.